requirements.txt
torch
torchvision
pillow
transformers
gradio
pillow
Wand
qdrant-client==1.8.0
